Split with a new crowdsfx.c
